refactor(v0): extract engine runner service from plan_session_service
